refactor: updated class ComputerRepository with INSERT method from Computer Class
