+++ Repositories/ComputerRepository.cs
@@ -12,6 +12,7 @@ class ComputerRepository
     {
         _databaseConfig = databaseConfig;
     }
+
     public List<Computer> GetAll() 
     {
         var computers = new List<Computer>(); 
@@ -33,5 +34,21 @@ class ComputerRepository
         connection.Close(); 
 
         return computers;
+    }
+
+    public void Save(Computer computer)
+    {
+        var connection = new SqliteConnection(_databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery();
+
+        connection.Close();
     }
 }
